Add `copy-video` function to video/core.
This function works like struct-copy, with the exception that it only works
on video objects. Furthermore, only video objects with video-ops methods.
(Which is all of the ones provided by this library). The upside
to using video-copy is that it does dynamic-dispatch for you, producing
the correct type of video.

It is important to note that this functionality is currently untested.
This is in part because I currently do not have the tests provided by
raco test working on my machine, but they do seem to be working via
travis.