A few more experiments.
I have tracked the memory leak down to the filtergraph. However, I
have not been able to stop the memory from exploding.

I know its a problem with how I'm using the filtergraph because:

1. Putting the compiled graph into the ffmpeg command line app makes
the issue go away,
2. I removed the filtergraph and the encoder (leaving the decoder in tact)
and the issue went away, and
3. I enabled the filtergraph (and just freed the output), but left the
encoder intact, and the issue came back.